✅(project) handle HTTP HEAD requests
Make LRS respond to any HTTP HEAD request as it would have responded to an otherwise identical HTTP GET request except the message-body MUST be omitted and the Content-Length header MAY be omitted.
Improve xapi specification conformance tested via lrs-conformance-test-suite :
- 3 tests that failed now pass,
- 0 test that passed now fail,
- a total of 600/1365 test passed.
