+++ tests/api/conftest.py
@@ -1,0 +1,1 @@
+from ..fixtures.statements import insert_statements_and_monkeypatch_backend  #noqa: F401
